+++ hopsworks-common/src/main/java/io/hops/hopsworks/common/featurestore/trainingdatasets/hopsfs/HopsfsTrainingDatasetController.java
@@ -16,22 +16,16 @@
 
 package io.hops.hopsworks.common.featurestore.trainingdatasets.hopsfs;
 
-import io.hops.hopsworks.persistence.entity.featurestore.storageconnector.hopsfs.FeaturestoreHopsfsConnector;
-import io.hops.hopsworks.persistence.entity.featurestore.trainingdataset.TrainingDataset;
-import io.hops.hopsworks.persistence.entity.featurestore.trainingdataset.hopsfs.HopsfsTrainingDataset;
-import io.hops.hopsworks.persistence.entity.hdfs.inode.Inode;
 import io.hops.hopsworks.common.featurestore.storageconnectors.FeaturestoreStorageConnectorType;
 import io.hops.hopsworks.common.featurestore.trainingdatasets.TrainingDatasetDTO;
 import io.hops.hopsworks.common.hdfs.inode.InodeController;
-import io.hops.hopsworks.common.featurestore.storageconnectors.hopsfs.FeaturestoreHopsfsConnectorFacade;
-import io.hops.hopsworks.exceptions.FeaturestoreException;
-import io.hops.hopsworks.restutils.RESTCodes;
+import io.hops.hopsworks.persistence.entity.featurestore.trainingdataset.TrainingDataset;
+import io.hops.hopsworks.persistence.entity.featurestore.trainingdataset.hopsfs.HopsfsTrainingDataset;
 
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
 import javax.ejb.TransactionAttribute;
 import javax.ejb.TransactionAttributeType;
-import java.util.logging.Level;
 
 /**
  * Class controlling the interaction with the hopsfs_training_dataset table and required business logic
@@ -40,48 +34,8 @@ import java.util.logging.Level;
 @TransactionAttribute(TransactionAttributeType.NEVER)
 public class HopsfsTrainingDatasetController {
   @EJB
-  private HopsfsTrainingDatasetFacade hopsfsTrainingDatasetFacade;
-  @EJB
-  private FeaturestoreHopsfsConnectorFacade featurestoreHopsfsConnectorFacade;
-  @EJB
   private InodeController inodeController;
 
-  /**
-   * Create and persiste a HopsFS training dataset
-   * @param connector
-   * @param inode
-   * @return
-   * @throws FeaturestoreException
-   */
-  @TransactionAttribute(TransactionAttributeType.REQUIRED)
-  public HopsfsTrainingDataset createHopsfsTrainingDataset(FeaturestoreHopsfsConnector connector, Inode inode) {
-    HopsfsTrainingDataset hopsfsTrainingDataset = new HopsfsTrainingDataset();
-    hopsfsTrainingDataset.setInode(inode);
-    hopsfsTrainingDataset.setFeaturestoreHopsfsConnector(connector);
-    hopsfsTrainingDatasetFacade.persist(hopsfsTrainingDataset);
-    return hopsfsTrainingDataset;
-  }
-
-  /**
-   * Verify hopsfsconnectorid
-   *
-   * @param hopsfsConnectorId the id to verify
-   * @throws FeaturestoreException
-   */
-  private FeaturestoreHopsfsConnector verifyHopsfsTrainingDatasetConnectorId(Integer hopsfsConnectorId)
-    throws FeaturestoreException {
-    if (hopsfsConnectorId == null) {
-      throw new IllegalArgumentException(RESTCodes.FeaturestoreErrorCode.HOPSFS_CONNECTOR_ID_NOT_PROVIDED.getMessage());
-    }
-    FeaturestoreHopsfsConnector featurestoreHopsfsConnector =
-      featurestoreHopsfsConnectorFacade.find(hopsfsConnectorId);
-    if (featurestoreHopsfsConnector == null) {
-      throw new FeaturestoreException(RESTCodes.FeaturestoreErrorCode.HOPSFS_CONNECTOR_NOT_FOUND, Level.FINE,
-        "HOPSFS connector with id: " + hopsfsConnectorId + " was not found");
-    }
-    return featurestoreHopsfsConnector;
-  }
-  
   /**
    * Converts a Hopsfs Training Dataset entity into a DTO representation
    *
